+++ Tests/SwiftLintFrameworkTests/RuleTests.swift
@@ -136,4 +136,6 @@ class RuleTests: XCTestCase {
     func testDifferentSeverityLevelRulesNotEqual() {
         XCTAssertFalse(RuleWithLevelsMock().isEqualTo(RuleWithLevelsMock2()))
     }
+    
+    
 }